Constrain MAX_PAYLOAD_LEN, refine parser checksum implementation
- Constrain MAX_PAYLOAD_LEN to the appropriate data type of u16
- Move Checksum to separate module and encapsulate checksum calculations more
- Improve documentation of Checksum calculator
- Make the Checksum calculator support streaming checksum calculations and a one shot calculation
- Add tests for the checksum calculator

+++ ublox/src/lib.rs
@@ -65,7 +65,7 @@ pub trait UbxProtocol: Send + Sized {
     type PacketRef<'a>: Into<UbxPacket<'a>>;
 
     /// The maximum payload length supported by this protocol version.
-    const MAX_PAYLOAD_LEN: usize;
+    const MAX_PAYLOAD_LEN: u16;
 
     /// Matches a Class ID, Message ID, and payload to a specific packet type.
     fn match_packet(
