+++ ublox/src/proto23.rs
@@ -20,7 +20,7 @@ pub struct Proto23;
 impl crate::UbxProtocol for Proto23 {
     type PacketRef<'a> = PacketRef<'a>;
 
-    const MAX_PAYLOAD_LEN: usize = packetref_proto23::MAX_PAYLOAD_LEN as usize;
+    const MAX_PAYLOAD_LEN: u16 = packetref_proto23::MAX_PAYLOAD_LEN;
 
     fn match_packet(
         class_id: u8,
