+++ ublox/src/proto27.rs
@@ -19,7 +19,7 @@ pub struct Proto27;
 
 impl crate::UbxProtocol for Proto27 {
     type PacketRef<'a> = PacketRef<'a>;
-    const MAX_PAYLOAD_LEN: usize = packetref_proto27::MAX_PAYLOAD_LEN as usize;
+    const MAX_PAYLOAD_LEN: u16 = packetref_proto27::MAX_PAYLOAD_LEN;
 
     fn match_packet(
         class_id: u8,
